Track toxic ramp stage; fix Reversal/Flail HP bucketing
Add a per-side badly-poisoned (tox) ramp-stage counter on the active mon
token (NUMERIC_TOXIC_STAGE = stage/15): a tox status starts it at 1, each
turn it escalates (the 1/16, 2/16, ... damage), and it resets on switch
(Gen 3). This is distinct from the status:tox categorical, which only
marks the type — the counter carries the escalating severity that drives
stall / switch timing. Tracked symmetrically for both sides.

Also fix resolve_move_base_power for Reversal/Flail: Gen 3 buckets the
floored value floor(48*curHP/maxHP), so compare the integer bucket rather
than the raw ratio (e.g. 3/96 -> 1.5 -> floor 1 -> 200 BP, not 150). Add
boundary-ratio tests and a Future-Sight-survives-switch test.

Spec numeric count 37 -> 38; doc updated.

+++ src/pokezero/dex.py
@@ -347,7 +347,8 @@ def resolve_move_base_power(move: "MoveInfo", user_hp_fraction: float | None = N
         return move.base_power
     fraction = max(0.0, min(1.0, user_hp_fraction))
     if move.id in _HP_BASE_POWER_LOW:
-        scaled = 48 * fraction
+        # Gen 3 buckets the floored value p = floor(48 * curHP / maxHP), not the raw ratio.
+        scaled = int(48 * fraction)
         if scaled <= 1:
             return 200
         if scaled <= 4:
